- Fixed potential issue with calling win32 SetWindowPos on 64-bit versions of windows - Cleanup

+++ OutlookDesktop/UnsafeNativeMethods.cs
@@ -49,7 +49,7 @@ namespace OutlookDesktop
         public static extern bool ReleaseCapture();
 
         [DllImport("user32.dll")]
-        public static extern IntPtr SendMessage(IntPtr hWnd, UInt32 Msg, Int32 wParam, Int32 lParam);
+        public static extern IntPtr SendMessage(IntPtr hWnd, UInt32 Msg, IntPtr wParam, IntPtr lParam);
 
         [DllImport("user32.dll", EntryPoint = "SetWindowPos")]
         [return: MarshalAs(UnmanagedType.Bool)]
